// src/app/swagger-api-generated/model/appealRes.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { LocationRes } from './locationRes';


export interface AppealRes {
    creationDate?: string;
    creator?: string;
    date?: string;
    file?: any;
    id?: number;
    level?: AppealRes.LevelEnum;
    location?: LocationRes;
    priority?: AppealRes.PriorityEnum;
    responseFile?: any;
    status?: AppealRes.StatusEnum;
    title?: string;
}
export namespace AppealRes {
    export type LevelEnum = 'FEDERAL' | 'LOCAL' | 'REGIONAL';
    export const LevelEnum = {
        FEDERAL: 'FEDERAL' as LevelEnum,
        LOCAL: 'LOCAL' as LevelEnum,
        REGIONAL: 'REGIONAL' as LevelEnum
    };
    export type PriorityEnum = 'HIGH' | 'LOW' | 'MEDIUM';
    export const PriorityEnum = {
        HIGH: 'HIGH' as PriorityEnum,
        LOW: 'LOW' as PriorityEnum,
        MEDIUM: 'MEDIUM' as PriorityEnum
    };
    export type StatusEnum = 'CLOSED' | 'OPENED' | 'WAIT';
    export const StatusEnum = {
        CLOSED: 'CLOSED' as StatusEnum,
        OPENED: 'OPENED' as StatusEnum,
        WAIT: 'WAIT' as StatusEnum
    };
}